// web/src/shared/api/generated/model/weakSpotsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { WeakLessonResult } from './weakLessonResult';
import type { WeakObjectiveResult } from './weakObjectiveResult';

export interface WeakSpotsResult {
  lessons: WeakLessonResult[];
  objectives: WeakObjectiveResult[];
}